Move position classes to seperate file
